+++ src/apps/molresponse/iterate_xy.cc
@@ -350,7 +350,7 @@ GammaResponseFunctions TDHF::ComputeGammaFunctions(
 
   std::vector<response_space> y_phi;
   std::vector<response_space> x_phi;
-  for (int b = 0; b < m; b++) {
+  for (size_t b = 0; b < m; b++) {
     y_phi.push_back(response_space(world, n, n));
     if (Rparams.omega != 0.0) {
       x_phi.push_back(response_space(world, n, n));
@@ -358,8 +358,8 @@ GammaResponseFunctions TDHF::ComputeGammaFunctions(
   }
 
   //
-  for (int b = 0; b < m; b++) {
-    for (int k = 0; k < n; k++) {
+  for (size_t b = 0; b < m; b++) {
+    for (size_t k = 0; k < n; k++) {
       // multiply the kth orbital to vector of y[b] response funtions...apply op
       // // to each product (TODO) //split apply and
       y_phi[b][k] = apply(world, op, mul(world, Gparams.orbitals[k], y[b]));
@@ -368,7 +368,7 @@ GammaResponseFunctions TDHF::ComputeGammaFunctions(
       }
     }
   }
-  for (int b = 0; b < m; b++) {
+  for (size_t b = 0; b < m; b++) {
     if (Rparams.omega != 0.0) {
       x_phi[b].truncate_rf();
     }
@@ -376,7 +376,7 @@ GammaResponseFunctions TDHF::ComputeGammaFunctions(
   }
   real_function_3d temp_J;
   // for each response state we compute the Gamma response functions
-  for (int b = 0; b < m; b++) {
+  for (size_t b = 0; b < m; b++) {
     // apply - save and truncate - multiply
     //
     //
@@ -389,7 +389,7 @@ GammaResponseFunctions TDHF::ComputeGammaFunctions(
       W[b] = mul(world, Wphi[b], Gparams.orbitals);
     }
     //  compute each of the k response parts...
-    for (int k = 0; k < n; k++) {
+    for (size_t k = 0; k < n; k++) {
       // Jcoulb*orbital[k]
       //
       Kx[b][k] = dot(world, x[b], phi_phi[k]);
